pyproject.toml
[project]
name = "receipt-buddy"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "google-api-python-client>=2.182.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "google-genai>=1.38.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pydantic>=2.11.9",
]
